+++ apps/rhf-mui-docs/src/constants/index.ts
@@ -1,5 +1,4 @@
 export { default as ExternalLinks } from './external-links';
-export { default as InternalLinks } from './internal-links';
 export { default as PropsDescription } from './props';